feat :: `GenerateQrCode` 유틸화를 통한 관심사 분리
